requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2019.9.11
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-filter==2.4.0
django-polymorphic==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jsonapi==4.1.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pipenv==2018.11.26
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==16.7.7
virtualenv-clone==0.5.3
zipp==3.4.1
zope.interface==5.2.0
